Other Services/FastAPI/requirements.txt
fastapi
uvicorn
python-dotenv
requests
clickhouse-connect
pydantic
